Rename and split components
