+++ madgui/__init__.py
@@ -1,0 +1,32 @@
+# encoding: utf-8
+"""
+MadGUI - interactive GUI application for MADX via cpymad.
+
+Usage:
+    madgui [-l <logs>]
+    madgui (--help | --version)
+
+Options:
+    -l <logs>, --log=<logs>     Set log directory (default is '~/.madgui')
+    -h, --help                  Show this help
+    -v, --version               Show version information
+
+Contact information:
+
+    Thomas Gläßle <t_glaessle@gmx.de>
+
+Website:
+
+    https://github.com/coldfix/madgui
+"""
+"""
+To start MadGUI type at your terminal:
+
+    python -m madgui
+
+or simply:
+
+    madgui
+"""
+
+__version__ = '0.2'
