+++ madgui/component/__init__.py
@@ -1,0 +1,11 @@
+# encoding: utf-8
+"""
+Plugin component package.
+
+This package contains controller components that are part of the main
+application. These are connected to the application via a named event
+mechanism.
+"""
+
+# force new style imports
+from __future__ import absolute_import
