+++ madgui/util/__init__.py
@@ -1,0 +1,9 @@
+# encoding: utf-8
+"""
+Package with various simple non-GUI utilities.
+
+The modules and objects defined in this package have no dependencies on
+other parts of madgui and are neither subscribers nor submitters of any
+events. They should also have little to no dependencies on non-standard
+3rd party modules.
+"""
